src/main/java/org/example/model/Product.java
package org.example.model;

import java.math.BigDecimal;
import java.util.Objects;

public class Product {
    private String id;
    private String productName;
    private Double price;

    public Product(String id, String productName, Double price) {
        this.id = id;
        setProductName(productName);
        setPrice(price);
    }


    public String getId() {
        return id;
    }

    public String getProductName() {
        return productName;
    }

    public void setProductName(String productName) {
        this.productName = productName;
    }

    public BigDecimal getPrice() {
        return BigDecimal.valueOf(price);
    }

    public void setPrice(Double price) {
        this.price = price;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Product product = (Product) o;
        return Objects.equals(id, product.id) && Objects.equals(productName, product.productName) && Objects.equals(price, product.price);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, productName, price);
    }
}
